fix: This project also has no actual dependencies
