--- .babelrc
@@ -1,9 +1,0 @@
-{
-  "presets": [
-    ["es2015", { "modules": false }]
-  ],
-  "plugins": [
-    "syntax-object-rest-spread",
-    "transform-object-rest-spread"
-  ]
-}
