+++ .gitignore
@@ -7,3 +7,4 @@ coverage
 .idea/
 *.iml
 *.swp
+grammar.js
