+++ index.js
@@ -1,1 +1,13 @@
-module.exports = require('./lib');
+const parser = require('./parser');
+
+module.exports = {
+  parse: parser.parse,
+  accepts: function(str) {
+    try {
+      parser.parse(str);
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
+};
