+++ package.json
@@ -12,9 +12,6 @@
   "dependency-check": {
     "entries": [
       "./index.js"
-    ],
-    "ignore": [
-      "debug"
     ]
   },
   "license": "Apache-2.0",
@@ -26,29 +23,20 @@
     "model"
   ],
   "scripts": {
-    "test": "mocha --require babel-core/register",
-    "compile": "npm run babel && npm run pegjs",
-    "postcompile": "babel ./lib/parser.js --out-file ./lib/parser.js",
-    "babel": "babel ./src --out-dir ./lib",
-    "pegjs": "pegjs -o ./lib/parser.js < ./src/grammar.pegjs",
-    "check": "mongodb-js-precommit ./src ./test",
+    "test": "mocha",
+    "compile": "npm run pegjs",
+    "pegjs": "pegjs -o ./parser.js < ./grammar.pegjs",
+    "check": "mongodb-js-precommit",
     "prepublishOnly": "npm run compile",
-    "pretest": "npm run compile"
+    "pretest": "npm run compile",
+    "precheck": "npm run compile"
   },
   "main": "index.js",
-  "dependencies": {
-    "bson": "^4.0.2",
-    "debug": "^4.1.1",
-    "lodash": "^4.17.15",
-    "pegjs": "^0.10.0"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-plugin-syntax-object-rest-spread": "^6.13.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
     "eslint-config-mongodb-js": "^5.0.3",
     "mocha": "^7.0.0",
-    "mongodb-js-precommit": "^2.0.0"
+    "mongodb-js-precommit": "^2.0.0",
+    "pegjs": "^0.10.0"
   }
 }
